pyproject.toml
[project]
name = "discord-steam-archival-bot"
version = "0.1.0"
description = "Discord bot to automatically archive Steam clips"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.13.2",
    "aiosqlite>=0.22.1",
    "discord-py>=2.6.4",
]
